Extended Isolation Forest - move IsolationTree test to separate class
